Remove unreachable exception in ShouldMatch
The member expressions came off the expected type, so there is no way for this to fail with a null-ref currently.
Setter-only properties will still be retrieved, but will throw an invalid operation on the .GetValue call later.
This seems ok however, as attempting to verify a set-only property is really an invalid test assert.