Prevent reflection/null ref errors when using inherited types with ShouldLookLike (#118)
* Prevent reflection/null ref errors when using inherited types with ShouldLookLike

This means comparing different levels of an object hierarchy doesn't explode messily. E.g., comparing an EntityFramework proxy to it's actual class.

* Add additional test proving behaviour aligns with ExpectedObject overloads

* Switch from assert failure to InvalidOperationException when property is not found

Technically this is a test setup failure, rather than an assertion that has failed, as it indicates incorrect types are being searched for.

* Remove unreachable exception in ShouldMatch

The member expressions came off the expected type, so there is no way for this to fail with a null-ref currently.
Setter-only properties will still be retrieved, but will throw an invalid operation on the .GetValue call later.
This seems ok however, as attempting to verify a set-only property is really an invalid test assert.